requirements.txt
absl-py==0.11.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.1
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
git+https://github.com/facebookresearch/fastText.git
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.5.0
gast==0.3.3
gensim==4.0.1
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imbalanced-learn==0.8.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
loguru==0.5.3
lxml==4.6.3
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
newspaper3k==0.2.8
nltk==3.6.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
pkg_resources==0.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycparser==2.20
pycurl==7.43.0.6
pydeck==0.6.2
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.8.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.1
scikit-learn==0.24.1
scikit-multilearn==0.2.0
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.7.1
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smart-open==5.0.0
smmap==4.0.0
soupsieve==2.2.1
SQLAlchemy==1.4.18
stop-words==2018.7.23
streamlit==0.85.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wordcloud==1.8.1
wrapt==1.12.1
